+++ Source/ARC.Client/Network/Packets/InboundConnectRequest.cs
@@ -5,7 +5,7 @@ public class InboundConnectRequest
 {
     public double ServerTime { get; }
     public ulong Cookie { get; }
-    public ushort ClientId { get; }
+    public uint ClientId { get; }
     public byte[] ServerSeed { get; }
     public byte[] ClientSeed { get; }
 
@@ -13,7 +13,7 @@ public class InboundConnectRequest
     {
         ServerTime = packet.DataReader.ReadDouble();
         Cookie = packet.DataReader.ReadUInt64();
-        ClientId = (ushort)packet.DataReader.ReadUInt32();
+        ClientId = packet.DataReader.ReadUInt32();
         ServerSeed = packet.DataReader.ReadBytes(4);
         ClientSeed = packet.DataReader.ReadBytes(4);
     }
